+++ event-log/src/main/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/MigrationRequestEvent.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.eventlog.subscriptions.tsmigration
+package io.renku.eventlog.subscriptions.tsmigrationrequest
 
 import cats.Show
 import cats.syntax.all._
