+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/SubscriberTrackerSpec.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.eventlog.subscriptions.tsmigration
+package io.renku.eventlog.subscriptions.tsmigrationrequest
 
 import Generators._
 import TypeSerializers._
@@ -24,7 +24,7 @@ import cats.data.Kleisli
 import cats.effect.IO
 import io.renku.db.SqlStatement
 import io.renku.eventlog.InMemoryEventLogDbSpec
-import io.renku.eventlog.subscriptions.tsmigration.MigrationStatus._
+import io.renku.eventlog.subscriptions.tsmigrationrequest.MigrationStatus._
 import io.renku.events.consumers.subscriptions.{SubscriberUrl, subscriberIds, subscriberUrls}
 import io.renku.generators.CommonGraphGenerators.serviceVersions
 import io.renku.generators.Generators.Implicits._
