+++ event-log/src/test/scala/io/renku/eventlog/subscriptions/tsmigrationrequest/SubscriptionPayloadDeserializerSpec.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.eventlog.subscriptions.tsmigration
+package io.renku.eventlog.subscriptions.tsmigrationrequest
 
 import Generators._
 import cats.syntax.all._
